Don't explode when the head repo has been deleted.
The GitHub API docs don't mention it, but if a pull request is created
then the head (as opposed to the base) repo is deleted, the head.repo
property is null/nil.

This change guards against that, and the test repos have been updated
with an example (pull request #4). Still no proper fixtures.